More robust TOS 2.06+ fix
We don't need the TIMEKPR patch anymore. Instead, we retrieve the
XBIOS jumptable directly using the "undefined opcode" trick. This
coerce the TOS to give us its jumptable. This is more robust because we
have more chance to get the original settime even if we're not the
first in the XBRA chain. And we now call settime directly instead
of "falling through" XBIOS.

The only restriction is that we could interfere with another program
redirecting settime.